Allow a shell command view to be refreshed by re-running the command.
The refresh is usually done after some command has finished running. For example if we had a shell command that output a list of files, and then another command that could delete a file from the list, we'd like the list of files to be updated each time a file is deleted.